move tests and formatter in irve analysis
